document-processor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
llama-index==0.10.0
llama-index-embeddings-ibm==0.1.0
llama-index-llms-ibm==0.1.0
llama-index-vector-stores-chroma==0.1.0
llama-index-readers-file==0.1.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.6
chromadb==0.3.29
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
textblob>=0.17.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
ibm-watson>=8.1.0
ibm-cloud-sdk-core>=3.18.0
